Move simplisafe base entity to separate module
