+++ homeassistant/components/simplisafe/button.py
@@ -15,8 +15,9 @@ from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import SimpliSafe, SimpliSafeEntity
+from . import SimpliSafe
 from .const import DOMAIN
+from .entity import SimpliSafeEntity
 from .typing import SystemType
