+++ homeassistant/components/simplisafe/const.py
@@ -13,5 +13,12 @@ ATTR_ENTRY_DELAY_AWAY = "entry_delay_away"
 ATTR_ENTRY_DELAY_HOME = "entry_delay_home"
 ATTR_EXIT_DELAY_AWAY = "exit_delay_away"
 ATTR_EXIT_DELAY_HOME = "exit_delay_home"
+ATTR_LAST_EVENT_INFO = "last_event_info"
+ATTR_LAST_EVENT_SENSOR_NAME = "last_event_sensor_name"
+ATTR_LAST_EVENT_SENSOR_TYPE = "last_event_sensor_type"
+ATTR_LAST_EVENT_TIMESTAMP = "last_event_timestamp"
 ATTR_LIGHT = "light"
+ATTR_SYSTEM_ID = "system_id"
 ATTR_VOICE_PROMPT_VOLUME = "voice_prompt_volume"
+
+DISPATCHER_TOPIC_WEBSOCKET_EVENT = "simplisafe_websocket_event_{0}"
